Consolidate api and data directories
This fixes unknown parent package import issues.
