+++ git-hooks/pre-commit
@@ -1,3 +1,3 @@
 docker-compose run --rm api poetry run autopep8 -v -a -r ./ --in-place
 docker-compose run --rm api poetry run flake8 .
-docker-compose run --rm api poetry run mypy api/ data/
+docker-compose run --rm api poetry run mypy api/
